feat: 학습로그 페이지와 메인페이지 분리
